docs/devel/9p: fix references to pictures
The original source https://wiki.qemu.org/Documentation/9p does not use
a prefixed path for pictures. The (auto converted) in-tree .rst version
of the 9p documentation though uses pictures that are placed into the
subdirectory 'img' for not mixing them with .rst files within the same
directory, so prefix references to those pictures with 'img/'.

Also the auto conversion tool 'pandoc' did not replace the image
references by a correct 'image' rst keyword, so correct that as well.

Signed-off-by: Christian Schoenebeck <qemu_oss@crudebyte.com>